ayvuapp/requirements/base.txt
# Django
asgiref==3.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-simple-history==3.0.0
Pillow==9.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0  # http://whitenoise.evans.io/en/stable/index.html
